Refactor aws repos to use common logic
